Introduction de slim
