+++ templates/footer.php
@@ -1,0 +1,17 @@
+      <hr>
+
+      <footer>
+        <p>&copy; payutc 2012-<?php echo date('Y') ?></p>
+      </footer>
+
+    </div> <!-- /container -->
+    <script src="js/jquery-1.7.2.min.js"></script>
+    <script src="js/bootstrap.js"></script>
+    <script type="text/javascript">
+            init();
+            selectPage(1);
+    </script>
+    <script src="js/virement.js"></script>
+    </script> 
+  </body>
+</html>
